feat: adds the relevant points detector package.
The ContourOp.java provides some functions needed in the other classes
for handle the contour and take some of his characteristics.
The Detector.java provides the functionalities for detect relevant
points.
The Linearizator.java provides the functionalities for calculate the
linearity and curves of a contour.
The SpaceScaleDetector provides the functionalities for detect relevant
points using the scale space.